Refactor import/export container into modular workflows
